# store_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/orders/{id}": {
            "get": {
                "description": "Подробная информация о заказе (только пренадлежащие пользователю)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Получить информацию по заказу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID заказа (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OrderInfo"
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Ошибка доступа",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Отсутствует заказ",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}/status": {
            "patch": {
                "description": "Меняет статус заказа на один из возможных ('pending', 'paid', 'delivered', 'cancelled', 'on_the_way')",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Заменить статус заказа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID заказа (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус заказа",
                        "name": "new_status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Orders"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Статус успешно изменен"
                    },
                    "400": {
                        "description": "Некорректные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Ошибка доступа",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Отсутствует заказ",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/stores/{id}/item-types": {
            "get": {
                "description": "Возвращает список типов товара по ID магазина",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Получить список типов товара",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID магазина (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ItemType"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Типы товаров не найдены",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/stores/{id}/items": {
            "get": {
                "description": "Возвращает список товаров по ID типа товара",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Получить список товаров",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID типа товара (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Item"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Товары не найдены",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/http_response.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Item": {
            "type": "object",
            "properties": {
                "card_img": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "description": "ID из таблицы store_item",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "types_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "ItemType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "OrderInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/OrderItemInfo"
                    }
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "OrderItemInfo": {
            "type": "object",
            "properties": {
                "card_img": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "Orders": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "http_response.ErrResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}